reorganize files for fewer changes and moving parts
