+++ test/snapshots/helpers/with-renderer.ts
