+++ test/snapshots/package.json
@@ -10,9 +10,9 @@
     "build": "sdk-commands build"
   },
   "devDependencies": {
-    "@dcl/js-runtime": "file:../../../packages/@dcl/js-runtime",
-    "@dcl/ecs": "file:../../../packages/@dcl/ecs",
-    "@dcl/sdk": "file:../../../packages/@dcl/sdk",
-    "@dcl/sdk-commands": "file:../../../packages/@dcl/sdk-commands"
+    "@dcl/js-runtime": "file:../../packages/@dcl/js-runtime",
+    "@dcl/ecs": "file:../../packages/@dcl/ecs",
+    "@dcl/sdk": "file:../../packages/@dcl/sdk",
+    "@dcl/sdk-commands": "file:../../packages/@dcl/sdk-commands"
   }
 }
