--- test/snapshots/production-bundles/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "baseUrl": "."
-  },
-  "extends": "@dcl/sdk/types/tsconfig.ecs7.json",
-  "compileOnSave": false,
-  "exclude": ["bin"]
-}
